calculatrice ok

+++ Calculatrice.Operations/Modulo.cs
@@ -6,9 +6,11 @@
         {
         }
 
-        public override int Calculer(int a, int b)
+        public override decimal Calculer(int a, int b)
         {
             return a % b;
         }
+
+
     }
 }
