+++ Calculatrice.Operations/Multiplication.cs
@@ -6,9 +6,10 @@
         {
         }
 
-        public override int Calculer(int a, int b)
+        public override decimal Calculer(int a, int b)
         {
             return a * b;
         }
+      
     }
 }
